Convert skeleton classes to xtend
I'm not so sure I made the best choice of language. It doesn't have
quite the level of syntactic sugar as coffee or coco, and compiles
heavier than I'd like (with helper libs), which at that point I might as
well have gone to the much more mature Scala. The shortcuts that it does
have seem fine though, and as long as it works with Android Annotations
and RoboGuice and so on, I'll be fairly happy.